Move multicast internals out of rig.h into multicast.h
This data (I think) was part of multicast V1, and doesn't seem to be used any
  more. Since neither V1 or V2 are enabled, there should be no external
  conflicts.
